third_party/WebKit/LayoutTests/fast/events/onerror-no-constructor-expected.txt

CONSOLE MESSAGE: line 8: FAIL
CONSOLE ERROR: line 20: Uncaught PASS if this is reported as an uncaught exception.
